+++ src/lt_l1.c
@@ -14,6 +14,26 @@
 #include "libtropic_common.h"
 #include "lt_l1_port_wrap.h"
 
+#ifdef LT_PRINT_SPI_DATA
+#include "stdio.h"
+#define SPI_DIR_MISO 0
+#define SPI_DIR_MOSI 1
+void print_hex_chunks(const uint8_t *data, uint8_t len, uint8_t dir)
+{
+    if ((!data) || (len == 0)) {
+        return;
+    }
+    printf("%s", dir ? "  >>  TX: " : "  <<  RX: ");
+    for (size_t i = 0; i < len; i++) {
+        printf("%02X ", data[i]);
+        if ((i + 1) % 32 == 0) {
+            printf("\n          ");
+        }
+    }
+    printf("\n");
+}
+#endif
+
 lt_ret_t lt_l1_read(lt_l2_state_t *s2, const uint32_t max_len, const uint32_t timeout)
 {
 #ifdef LIBT_DEBUG
@@ -88,6 +108,9 @@ lt_ret_t lt_l1_read(lt_l2_state_t *s2, const uint32_t max_len, const uint32_t ti
                 return LT_L1_SPI_ERROR;
             }
             lt_l1_spi_csn_high(s2);
+
+            print_hex_chunks(s2->buff, s2->buff[2] + 4, SPI_DIR_MISO);
+
             return LT_OK;
 
             // Chip status does not contain any special mode bit and also is not ready,
@@ -132,6 +155,8 @@ lt_ret_t lt_l1_write(lt_l2_state_t *s2, const uint16_t len, const uint32_t timeo
 #endif
 
     lt_l1_spi_csn_low(s2);
+
+    print_hex_chunks(s2->buff, len, SPI_DIR_MOSI);
 
     if (lt_l1_spi_transfer(s2, 0, len, timeout) != LT_OK) {
         lt_l1_spi_csn_high(s2);